refactor(server): extract read-only data routes (4/8 server split)
Moves the 5 unauthenticated, read-only endpoints to their own Blueprint:
- GET /              (self-describing root, auto-discovers url_map)
- GET /ping          (keepalive)
- GET /api/data      (full JSON cache for dashboard)
- GET /api/health    (uptime + cycle counters)
- GET /api/stats     (quick DB stats)

These share a contract — no auth, no writes, served as the public read
surface — so they group naturally.

server.py change:
  - Removed all 5 inline route handlers (~50 LOC)
  - Added: from routes.data_routes import data_bp + app.register_blueprint
  - Net shrink: 46 LOC (767 → 721)

routes/data_routes.py:
  - 81 LOC, no logic change from the originals
  - Uses flask.current_app inside the / route's url_map iteration, avoids
    circular-importing the app object from server.
  - Imports state, build_cache, DB_PATH, get_stats from core/storage modules
    rather than server, breaking the implicit dependency on server's
    module namespace.

Tests: 219 passed (unchanged).
